Restructure crates
