--- atspi-client/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "atspi-client"
-version = "0.1.0"
-edition = "2021"
-description = "A basic AT-SPI client library."
-readme = "README.md"
-keywords = ["a11y", "linux"]
-categories = ["Accessibility", "Linux"]
-license = "Apache-2.0 OR MIT" # For ease of integration in the Rust ecosystem.
-repository = "https://github.com/odilia-app/atspi"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-default = ["async-io"]
-async-io = ["atspi-types/zbus-async-io", "atspi/async-io"]
-tokio = ["atspi-types/zbus-tokio", "atspi/tokio"]
-
-[dependencies]
-futures-lite = "1.13.0"
-zbus = "3.12.0"
-atspi-types = { path = "../atspi-types", default-features = false }
-atspi = { path = "../atspi", default-features = false, features = ["unstable-traits"] }
-tracing = "0.1.37"
-
-[dev-dependencies]
-rename-item = "0.1.0"
-tokio = { version = "1.28.1", features = ["macros", "rt-multi-thread"] }
-tokio-test = "0.4.2"
-enumflags2 = "0.7.7"
-async-std = "1.12.0"
