+++ atspi-common/src/error.rs
