+++ atspi-common/src/macros.rs
