+++ atspi/examples/focused-tokio.rs
@@ -1,10 +1,14 @@
+#[cfg(feature = "connection")]
 use atspi::events::object::{ObjectEvents, StateChangedEvent};
-use futures_lite::stream::StreamExt;
+#[cfg(feature = "connection")]
 use std::error::Error;
+#[cfg(feature = "connection")]
+use tokio_stream::StreamExt;
 
-#[async_std::main]
+#[cfg(feature = "connection")]
+#[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
-	let atspi = atspi_client::AccessibilityConnection::open().await?;
+	let atspi = atspi::AccessibilityConnection::open().await?;
 	atspi.register_event::<ObjectEvents>().await?;
 
 	let events = atspi.event_stream();
@@ -19,4 +23,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
 		}
 	}
 	Ok(())
+}
+#[cfg(not(feature = "connection"))]
+fn main() {
+  println!("This test can not be run without the \"connection\" feature.");
 }
