+++ atspi/src/cache.rs
@@ -5,7 +5,7 @@
 //!
 
 use crate::atspi_proxy;
-use atspi_types::CacheItem;
+use atspi_common::CacheItem;
 use zbus::zvariant::OwnedObjectPath;
 
 #[atspi_proxy(interface = "org.a11y.atspi.Cache", default_path = "/org/a11y/atspi/cache")]
